requirements.txt
matplotlib
pandas
sklearn
tensorflow
grpcio
protobuf==3.20.*
